# openapi-gpt.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AliExpress Affiliate API Proxy",
    "description": "Production-ready Python service for AliExpress Affiliate API with comprehensive product search, category management, and affiliate link generation capabilities. Perfect for custom GPT integration.",
    "version": "1.0.0",
    "contact": {
      "name": "AliExpress API Support",
      "url": "https://github.com/your-username/aliexpress-python-refactor"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://your-deployment-url.vercel.app",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/affiliate/links": {
      "post": {
        "summary": "Generate Affiliate Links",
        "description": "Generate affiliate tracking links for product URLs",
        "operationId": "generateAffiliateLinks",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AffiliateLinkRequest"
              },
              "example": {
                "urls": [
                  "https://www.aliexpress.com/item/1005001234567890.html",
                  "https://www.aliexpress.com/item/1005001234567891.html"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated affiliate links",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffiliateLinkResponse"
                },
                "example": {
                  "success": true,
                  "data": [
                    {
                      "original_url": "https://www.aliexpress.com/item/1005001234567890.html",
                      "affiliate_url": "https://s.click.aliexpress.com/e/_tracking_link_1",
                      "tracking_id": "gpt_chat",
                      "commission_rate": "30%"
                    }
                  ],
                  "metadata": {
                    "total_links": 1,
                    "processing_time_ms": 200
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/affiliate/link": {
      "get": {
        "summary": "Generate Single Affiliate Link",
        "description": "Generate an affiliate link for a single product URL",
        "operationId": "generateSingleAffiliateLink",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "description": "Product URL to convert to affiliate link",
            "schema": {
              "type": "string",
              "format": "uri",
              "example": "https://www.aliexpress.com/item/1005001234567890.html"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Generated affiliate link",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleAffiliateLinkResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AffiliateLinkRequest": {
        "type": "object",
        "properties": {
          "urls": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            },
            "maxItems": 50,
            "description": "List of product URLs to convert (max 50)",
            "example": [
              "https://www.aliexpress.com/item/1005001234567890.html"
            ]
          }
        },
        "required": ["urls"]
      },
      "AffiliateLink": {
        "type": "object",
        "properties": {
          "original_url": {
            "type": "string",
            "format": "uri",
            "description": "Original product URL",
            "example": "https://www.aliexpress.com/item/1005001234567890.html"
          },
          "affiliate_url": {
            "type": "string",
            "format": "uri",
            "description": "Generated affiliate URL",
            "example": "https://s.click.aliexpress.com/e/_tracking_link_1"
          },
          "tracking_id": {
            "type": "string",
            "description": "Affiliate tracking ID",
            "example": "gpt_chat"
          },
          "commission_rate": {
            "type": "string",
            "description": "Commission rate for this product",
            "example": "30%"
          }
        },
        "required": ["original_url", "affiliate_url", "tracking_id"]
      },
      "AffiliateLinkResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AffiliateLink"
            }
          },
          "metadata": {
            "type": "object",
            "properties": {
              "total_links": {
                "type": "integer",
                "example": 1
              },
              "processing_time_ms": {
                "type": "number",
                "example": 200
              }
            }
          }
        },
        "required": ["success", "data"]
      },
      "SingleAffiliateLinkResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/AffiliateLink"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "processing_time_ms": {
                "type": "number",
                "example": 150
              }
            }
          }
        },
        "required": ["success", "data"]
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Optional API key for enhanced access"
      }
    }
  },
  "security": [
    {},
    {
      "ApiKeyAuth": []
    }
  ],
...
}